Day19 in Rust, part 1 fixed
Switched Day19 to do a two part BFS.  We do an outer-level BFS on all key nodes, using a secondary BFS to locate the keys and their step distance.

Outer-BFS needs to use a priorty queue to find the minimal answer.  We also need to add reasonable de-duping on (position, keys) to make the search space not explode.